QA pass: 294-test suite, fix 6 findings, fix app compile error
- Add QA suites: golden corpus (44 utterances), cleanup property tests,
  state machine fuzzing, adversarial autocorrect, installer edge cases
- Fix VB-QA-01: number-like words exempt from stutter collapse (phone
  numbers survive dictation)
- Fix VB-QA-02: interrogative utterances get '?' terminal punctuation
- Fix VB-QA-03: 'actually no' self-correction marker
- Fix VB-QA-06: internal capitals (iPhone, VBoard) gate autocorrect
- Fix VB-QA-07: per-pack mutex prevents concurrent-install corruption
- Fix VB-QA-09: typed literal always kept in the suggestion strip
- Fix Double->Float mismatch in ModelDownloadService (CI compile error)
- docs/QA_REPORT.md: findings, traceability, 20-case manual test plan

+++ app/src/main/kotlin/com/vboard/app/models/ModelDownloadService.kt
@@ -88,7 +88,7 @@ class ModelDownloadService : LifecycleService() {
             val result = app.packInstaller.install(pack) { state ->
                 publish(packId, state)
                 if (state is PackState.Downloading) {
-                    updateNotification(pack.displayName, state.fraction)
+                    updateNotification(pack.displayName, state.fraction.toFloat())
                 }
             }
             publish(packId, result)
